fix: move sections, answers, and questions fetches to layout.ts to decrease loading time
